+++ src/doc-mirror/index.ts
@@ -37,7 +37,7 @@ export class DocumentModel implements EditableModel {
                     this.document.selection = options.selection;
                 }
                 if (!options.skipFormat) {
-                    return formatter.formatPosition(editor, true, {
+                    return formatter.formatPosition(editor, false, {
                         "format-depth": options.formatDepth ? options.formatDepth : 1
                      });
                 }